# api.yaml
...
openapi: 3.0.3
info:
  version: 1.0.0
  title: Grandma and friends's recipes
  description: The Recipes Website API provides developers with access to a comprehensive
    collection of recipes and related information from a recipes website. It allows
    developers to integrate recipe data into their own applications, websites, or
    services, enabling users to search for recipes, retrieve detailed recipe information,
    and perform various recipe-related actions.
paths:
  /login:
    post:
      tags:
      - User
      summary: the user authentication with the server
      description: user gives it's username and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/login"
      responses:
        '200':
          description: OK
...
  /register:
    post:
      tags:
      - User
      summary: new user added to the website
      description: current person gives it's detials to create a new user in the system
        with a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userDetails"
      responses:
        '200':
          description: Successful registration
        '400':
          description: Invalid request - the username is already taken or the passwords
            do not match or the country doesn't exist
        '500':
          description: Internal server error - the registration could not be completed
...
components:
  schemas:
...
    userDetails:
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 8
          pattern: "^[A-Za-z]+$"
          example: sonGoku
          description: The username for the new user
        firstName:
          type: string
          description: The first name of the new user
          example: goku
        lastName:
          type: string
          description: The last name of the new user
          example: son
        country:
          type: string
          description: The country of the new user
          example: Israel
        password:
          type: string
          minLength: 5
          maxLength: 10
          pattern: "^(?=.*[0-9])(?=.*[!@#$%^&*])[a-zA-Z0-9!@#$%^&*]+$"
          description: The password for the new user
          example: kamehameha
        confirmPassword:
          type: string
          description: Confirm the password for the new user
          example: kamehameha
        email:
          type: string
          format: email
          description: The email for the new user
      required:
      - username
      - firstName
      - lastName
      - country
      - password
      - confirmPassword
      - email
    login:
      properties:
        username:
          type: string
          description: The user's username
          example: sonGoku
        password:
          type: string
          description: The user's password
          example: kamehameha
      required:
      - username
      - password
...
servers:
  - description: Online Server
    url: https://aflalo-teva.cs.bgu.ac.il/